Switch to using Middleman
This commit switches the framework behind the static site
generation from Jekyll to Middleman.

Jekyll can't do many dynamic things. This means that we have to
generate a lot of data in rake tasks before we can start rendering
pages. In #10 we've seen that this necessitates creating thousands of
lines of pre-generated HTML in git.

Middleman is easier because we can build this site almost like a normal
rails application. We use helpers to pass an object to the template,
which we can render with ERB. It's also nice that we're using the same
thing as https://github.com/alphagov/tech-docs-template.