Redesign dashboard for applications
This commit updates the "dashboard" view. It uses the
tech-docs-template to show a table of contents on the left hand side.
This should make it easier to understand the structure. It merges the
two pages (previously "Applications" and "Operations").

When this page is first loaded, the `Dashboard` class will use the
GitHub API to fetch all alphagov repos (5 requests) and external repos.

To avoid having to do this on each page refresh in development we use
`Faraday::HttpCache`. This makes developing the dashboard really fast.