Add docs for publishing-api
This adds the documentation for the publishing-api docs.

The list of pages to import is hard coded. I've considered using the
GitHub API to fetch everything in the `/doc` directory. The downside to
that is that if pages are renamed, we'll inadvertently break the pages
here. Having a local list here allows us to set up redirects.