Add document type pages
This adds pages for all document types in the search index. We've
recently started indexing these, so not all pages in the index have
document types yet.

The user need for these pages is that we'd like to start creating a
canonical list of document types. This will allow us to make groupings
(like "publication" or "announcement") consistently across the site.