Documentation for the content-store api.
This is intended as the start of documenting the data we expose
externally via the public /api/content endpoint; putting it in the
developer docs for now for want of a better place.

Obviously there is some overlap with the schemas documentation which
already exists, but the audience is different.