Fix issue with in-document links
Documents imported from the publishing-api/docs folder have a
table-of-contents which doesn't work here.

The documents are published on github, and made available here.
Due to a mismatch in Markdown -> HTML rendering between Github and
Middleman/Redcarpet, generated fragment IDs are created using a
different algorithm.

In this PR, I override the default Redcarpet `header` function to
make it use a Github flavoured ID scheme.