Simpler system for linking to GitHub
This moves the responsibility to generate the GitHub links to the
template that's used by all pages. It now generates a link to the
source file if a `source_url` hasn't been provided, so that all pages
will automatically have an view-source link. For simplicity there's now
just one link.