Import non-sensitive pages from the opsmanual
Yesterday we did an audit of all opsmanual documents to determine
whether they are safe to make public, or need work. The pages are
listed in [this google spreadsheet][gs].

This imports the non-sensitive documents from the opsmanual. This
was done by first converting any remaining `.rst` files in the
opsmanual to markdown. A script was run to then import the pages.

We haven't touched the content of the pages apart from adding a warning
that is needs to be reviewed and some link updating.

[gs]:
https://docs.google.com/spreadsheets/d/1thQHw_SZyAuMCyBTnVuM_7_EBda8DTRT
pHZdL0y0eV0/edit#gid=0