Add review dates for all pages in manual
This sets a review date for all pages in the manual. The review date is
a random date between now and October 2017. Because there are 130 pages
in the manual we will need to review on average 5 pages per week. This
is an arbitrary picked number, but it feels like it should be doable.