Disable the in-page navigation
This disables the automatic updating of the URL with the current
fragment (/index.html#something) and the automatic highlighting in the
sidebar. The in-page nav was designed for large pages like the PaaS
docs (https://docs.cloud.service.gov.uk).

I've found that because we have very small pages, the URL updating is a
bit distracting and doesn't add much. It also auto-updates on load,
which will change the URL immediately. I've found myself removing the
fragment a lot when sharing a particular page.

Note that this doesn't prevent people from sharing specific headings,
that's still possible.