Add note about nfs permissions error on macOS Sierra
There's a bug in the nfs implementation of macOS Sierra which means we will get permissions errors when running govuk_puppet.  Advice in https://github.com/mitchellh/vagrant/issues/8061 is to run `ls -alR > /dev/null` on the host before running the VM to refresh the cache.  Hopefully a future macOS Sierra update will fix this.