Promote redis flavoured emergency banner docs
We have deprecated the old method of deploying emergency publishing
banners. We had two sets of docs for a while so we could compare te old
and new process and iterate the new improved emergency publishing.

A previous commit removed the old documentation and this one promotes
the new Redis flavoured version to the old url since it is now the
canonical version.