Fix path of HMRC PAYE app manifest in cache purge instructions
When a new version of the HMRC PAYE tool is uploaded, the XML manifest
file is updated and the Fastly cache is purged so old versions of the
app pick up the new version.

This commit updates the cache purge instructions to clear the cache for
the appropriate manifest file (the test one or the final real one) after
each has been uploaded.