Update Emergency Publishing docs
The page mentioned that the types of events would be detailed below. The information was half way down the page, in the section for cache busting. It's useful to know about these event types at the beginning of the process, so this commit copies that information into the prerequisites sections. Also remove some redundant wording.