Update elasticsearch backup guidance
I want to make sure we can follow this guidance for managing GOV.UK
search (the api-elasticsearch cluster) so I've assigned our team as
owner.

I think this is still out of date for logs-elasticsearch because we now
store snapshots on s3 using
https://www.elastic.co/guide/en/elasticsearch/reference/1.7/modules-snapshots.html
but I've left it out of this PR.

Review this in 3 months because we're likely to setup s3 backups for
api-elasticsearch too at some point.