Replace slave to backup mysql command
This changes the advice for copying backups between servers from copying
private keys to one of the servers to instead use agent forwarding.

Whilst agent forwarding is not ideal ([see this blog for
details](https://lyte.id.au/2012/03/19/ssh-agent-forwarding-is-a-bug/))
it is a strict improvement on copying private keys to a remote server
and having to remember to remove them.