Add tips on dealing with a broken bowl
Sometimes the VM ends up created without a fully working `bowl` command
so the advice in the getting started guide about running apps seems like
it is out of date.  The resolution is to manually install the bowler gem
rather than trying to use `bundle exec bowl` or ignoring `bowl` in favour
of using `startup.sh` in each app folder directly.

It's not clear to me why some VMs get into this state, and others don't,
but it's happened enough that some docs on solving it wouldn't go amiss.

[Fixes #424]