Review search docs
The cluster health one looks good, I've just updated a reference to
api-elasticsearch, because we'll shortly be replacing it with
rummager-elasticsearch, which runs a newer version.

I've linked the elasticsearch backup/restore to traffic replay.  We're
currently changing the indexing process so this is likely to change -
I've shortened the review time to reflect that.