Add frontend memcache clear step and echo env instructions
We've added a step to check your environment, hopefully this will help
people in a high stress situation to double check they're in the right environment.

We've also added a step to clear the memcache for the frontend class since government-frontend
now uses Dalli (memcache based) for it's caching, enabling a quicker update.

There is also a note left for the next 2ndline to see if a government-frontend restart is still required.