Add section on transition to dashboard
`bouncer` and `transition` are apps that we use to support sites
transitioning to gov.uk, but there are a handful of other tools we use as
aprt of this process.  We collect them all here in a new "Transition"
section on the dashboard.  The repos were found via [the
`govuk-transition` topic on github][govuk-transition].  Some of the repos
we link to are private, so they have to be added as sites.

[govuk-transition]: https://github.com/search?q=topic%3Agovuk-transition+org%3Aalphagov&type=Repositories