Use commit SHA to reference line in code
The Host model in Transition has been modified since this documentation was first written, which means that the link that references a particular line in the code is not longer highlighting the correct part. We should use a particular point in time to ensure that this doesn't happen in the future.