Add Algolia DocSearch
This adds [Algolia DocSearch][al]. It's a free service used by
prominent projects such as React, Stripe and Scala. We have an account
set up, which is configured in
https://github.com/algolia/docsearch-configs/blob/master/configs/publish
ing_service_gov_uk.json.

We previously discussed this in
https://github.com/alphagov/govuk-developer-docs/pull/128. We didn't go
for this then because there were questions about accessibility. I think
it would make sense to implement it now after all because the Google is
very slow, causing new pages to be indexed weeks after first being
published.

[al]: https://community.algolia.com/docsearch