Add rendering apps to document types & CSVs
This adds the rendering app of a document type to the document type
page and to the CSV. We'll use this to check which applications we have
to update for the first iterations in the navigation team.

The data is fetched from production and copied here because there is no
API available that can do this accurately. The search API contains most
of the information, but there's some content missing or incorrect.

There are some interesting issues in data (coming_soon being rendered
by both whitehall-frontend and government-frontend).