Add error description for no matches found
When using zsh and trying to work with the eyaml rake tasks, a common
error is one of `zsh: no matches found`. Although there is already a
note in the docs to use the `noglob` option when running zsh, the
associated error is not mentioned.

Add a description of the error to the Troubleshooting section so that
searching for the error text `zsh: no matches found:
eyaml:edit[integration]` in the docs returns a match and offers a
solution.