Update ssh tunnel and local dev vm docs for sidekiq-monitoring
The sidekiq-monitoring app now listens on port 3211 instead of port 80
so we need to change the port we tunnel to and from.  The dev VM now can
host the sidekiq-monitoring app directly without needing a tunnel so we
can remove that section entirely.

See: https://github.com/alphagov/govuk-puppet/pull/6981 for the changes
that support documentation.