Add a guide about the concept of stacks
The concept of stacks is something which is fundamentally different in
the AWS architecture. This is a small, high level view on what their use
case is, and how they may be used.

The ADR needs to be updated to present a clearer view on the
implementation of the stacks and how DNS works with them.