Update "Edit an existing route in the Router"
Drop the review time, as I'm not sure when production will move to
AWS, and thus a different way of accessing the relevant machine will
be needed.

Add a paragraph to make it clearer when manually editing routes is
appropriate, and that it's only a short term fix/change.

Also, improve some capitalisation.