Update "Assets: how they work" documentation
I've reduced `review_in` from 6 to 1 month, because of the ongoing work
on asset management.

I've changed the `owner_slack` from 2ndline to asset-management, because
this is the channel which the development team is currrently using.

I've added whitehall to the list of `related_applications`, because it
still currently includes separate asset management functionality as
explained in the body of the page.