Add content schema keywords
The Modelling Services team have been working on the content schema for step by step navigtion (previously called `task lists`).

We had to do a bit of research to find out how to represent certain properties of our, and thought it would be good to add to the Developer docs.