Update docs for SSHing to environment
It's always a good idea to abstract the technical detail into a wrapper
coded in bash, so add documentation about how to use this particular
wrapper.

This documents how to use the `govukcli` tool which will hopefully make
SSHing into different environments easier. This also allows the use of
one method to use in different environments while we migrate.