Add ValeLint to CI
I added Vale as a pre-commit hook in b3dcb10783b238de3fbc0acc808da2647ee851f7.

The idea is that we can "lint" our documentation for grammar and
readability. Documentation is often as important as the code we have to
support, so we should ensure it is concise and accessible, particularly
to people where English is not their first language.

This adds the tool to the CI chain. It can be further configured using
`.vale`, and further styles may be added to `styles/`.