Improves docs for user access alert
New joiners will trigger an alert in Icinga for "Check that correct users have access" which might make people on 2ndline remove their access. This alert happens when people haven't had a chance to open a PR in govuk-puppet to give themselves access, but have already been added to the alphagov organisation and can see repos.