New layout for the manual index page
The new tech docs template doesn't have a layout without a sidebar.

This implements a new simpler layout design with the sections in the
sidebar and the page titles in the main content area.

This is also an opportunity to merge the `Manual` and `ManualIndexPage`
classes because they do mostly the same thing.