How to monitor rabbitmq on integration/dev
The previous command doesn't work any more since we moved to AWS.

The ssh-fu is based on what we suggest in the sidekiq monitoring page.
If there's a simpler way to do this please let me know and I'll
update all the examples.

The user journey is still broken because you can't get the password
for integration without getting full production access.
Devs do not get access to govuk-secrets on day 1.