Update data.gov.uk docs with Heroku and Signon
The new Heroku stuff is currently acting as an integration environment,
so it also made sense to tidy up references to 'staging' and PaaS
services, so that the architecture docs are more environment agnostic -
the PaaS docs should be sufficient for people to discover PaaS-specific
information.

The Signon doc will need to evolve as we refine our organisation mapping
strategy and permission model.