A guide to removing users from Puppet
This is an attempt to clarrify the process that should be used as
different developers seem to approach this in different ways and if you
log onto any of our machines you'll see many home directories for staff
that have left.

A better alternative would be that the govuk_user module removed every
home directory of users who aren't available as this would simplify
the process.