Start documentation on how to support CKAN
As we are now relying on CKAN for metadata harvesting, this document
describes how to run various commands to interact with it.

The hosting setup is yet to be determined, and so this section has been
left blank.  Once hosting is confirmed the application_template can be
reverted to remove the pending_hosting? check.