Enforce use of classes for fabric tasks
Previously some fabric tasks would automatically set the group of
hosts on which they should run, such as 'cache' or 'whitehall_backend'.
The code to support this magic was causing all fabric tasks to take a
long time to run. As only a few tasks had this behaviour, it's more
consistent to enforce the group of hosts be specified when running all
fabric tasks, while also making it easy to remove the task bottleneck.