Rewrite doc for new rails app to be more accurate
The instructions for creating a rails app are now explicit in this
document, as using the previous automated scripts proved to be to
brittle due to rails changing over time.

Setting up a rails app involves lots of interacting services. Rather
than documenting the processes here, we've made an effort to improve the
surrounding docs and link to them from here.