Update CDN logs documentation
This updates the information on our logs to be up to date, we no longer
have a transition-logs-1 box and we send so much logs to S3 in different
scenarios that this feels something that isn't worth explaining in full. It
now includes a link to the new page on querying the CDN logs.