Add suggestion to restart the computer
Here we suggest to use OpenConnect, if we have troubles with Cisco AnyConnect,
but when I followed this guide, the command suggested didn't initially solve the
problem. The issue was fixed by restarting the computer and trying the same
command again, without launching Cisco AnyConnect first, so I'm adding this step
to the guide.