Spell "license" correctly
Every time I create a README file I mess up this link to the license.
I thought I just couldn't spell, but it turns out this manual is
sabotaging me every time.

It's spelled with an S in the text but a C in the link.

The correct spelling in British english is licence when you use it
as a noun and license when you use it as a verb.
Source: https://www.grammarly.com/blog/licence-license/

I also put a .txt on the end because lots of our repos use this
extension.