Specify where applications are hosted
This adds where the application is hosted to the application page,
application index pages and CSV export. It should make it easier to
find the location of the apps, as well as automating some parts of the
docs. Source of the apps is [this Trello ticket on 2nd
line](https://trello.com/c/49OGPxEk/314-%F0%9F%94%AD-where-are-our-apps-
apps-are-moving-from-carrenza-to-aws).

This also fixes the SSH instructions for apps hosted in Carrenza and
AWS.