Use search, expiry feature and layouts from gem
This is a bit of a big commit, but it should be fine since it's mostly
deletions and it's all interconnected.

Things that happen in this commit:

- We remove the layout overrides that exist in these docs, in favour of
the layouts provided by the gem.
- The layouts now provide the meta tags functionality
(https://github.com/alphagov/tech-docs-gem/pull/11), so we can remove
our own code for that.
- The page review is now provided by the gem as well
(https://github.com/alphagov/tech-docs-gem/pull/12), so we get rid of
the custom code for that too.
- We turn on the search functionality from the gem which was added in
https://github.com/alphagov/tech-docs-gem/pull/28 and follow up PRs.

After this commit, we're mostly running code from the gem instead of
our own. This should make it easier to update, and makes it more likely
that we'll contribute back to the gem.