Add how to acknowledge emails where subject doesn’t match title
If the title of a content item changes after it was first published, but before the email alerts has been sent, the email alert
check will fail. Once you’ve confirmed  that the email was in fact sent, it’s ok to add it to the list of emails for the check to ignore.