Update shell commands to be more readable
Some of the commands listed in the readme were wrapped in single backticks, which meant that the commands were displayed inline. This made it harder to read as two commands would appear on the same line next to each other, rather than on their own lines. Wrapping the code in triple backticks makes it a block, with the commands on individual lines.