Describes storing commands needed between VM destroys
Occasionally, you just cannot avoid running `vagrant destroy` on a broken VM.

When that happens, you need to set up a few things from scratch:

- Git global configuration
- Symlink for `govukcli`
- Any permissions fixing / `chown`ing required
- Any custom aliases or theming you like to apply to your terminal

Encouraging the idea of 'throwaway VMs', we should make it easier to get a VM
back up and running after a destroy. This is a good first step for that.