Document redirect 404 issue and workaround
In Zendesk ticket 3712180 I had to process a redirect. I followed these instructions, initially
on integration, and they worked perfectly. When I followed the same instructions on production,
the redirect was applied but the new URL would 404.

Working with @cbaines we saw that the `downstream_low` queue in publishing-api was quite full,
and suspect that the rake task had put it in the downstream_low queue. In the meantime we ran
a rake task to bump it to high priority to process the content change more quickly. I feel it is worth
documenting the workaround in the same instructions that led to the issue.