Pulls in content-store README
In order to keep the information in one place and reduce duplication,
and to keep the documentation closer to the code, I've followed the
style of the [step by step docs](https://github.com/alphagov/govuk-developer-docs/blob/master/source/manual/side-by-side-browser.html.md.erb)
and pull in an external README instead.

I've decided to keep the 'limitations' part separate for the time being
as it covers areas outside of the content-store itself and provides a
bit of context more than anything.