Explain GOV.UK architecture
Pieces together as much knowledge as possible, explaining where
and how things happen at different points in a request, publish
or deployment.

We cover this in irregular "GOV.UK Architecture" sessions that
rely on seniors / leads to run them, but inevitably things get
missed or lack detail, and anyone who misses the session misses
out on that knowledge.

The aim of this document is to give a decent overview to every
part of the stack, in a way that won't fall badly out of date too
quickly.