Updates dependencies using 'bundle update'
I had to manually specify a minimum version of middleman:

    `gem 'middleman', '~> 4.3.4'`

Otherwise for some reason it would install `4.3.0.rc.4`, which
has a stricter `YAML.load_file` implementation that requires
converting Date and Symbol to string in all of our YAML files.

I also had to manually specify a _maximum_ version of html-pipeline:

    `gem 'html-pipeline', '~> 2.5.0'`

Because it was causing issues with ASCII/UTF-8 processing which is
preventing the docs from building. @timblair had the same issue in
April:

https://github.com/alphagov/govuk-developer-docs/commit/e9541eb4d01d1fb4b883041ca211613086618641

Specifying the version in Gemfile should stop this from creeping up
again.