Fix tests by fixing file format & updating fixtures + url
There was an 'allow-document-types-fixture.json' file which contained YAML
and which was referred to as YAML by the tests. Older versions of YAML loading
libs were more forgiving of this, but the latest version we've updated to is
more strict, causing the tests to fail. I've renamed the fixture to .yml.

Next I updated the URL in document_types_csv_spec and document_types_spec to
match the URL called in document_types.rb, as this is the URL we're stubbing.

I had to remove 'announcement' from the fixture data, as that is no longer a
valid document_type. But it may be important to test that applications with
no dependencies return "" in their CSV, so I found another example of an app
that has no dependencies ('redirect') and test that instead.