Fix links to jenkins jobs to point to AWS instances
Some of our jobs now run in our AWS infrastructure. Point our links
towards the AWS jobs so that they correctly reflect our current
architecture.

Remove some redundant links and consolidate some duplicates.

Move the link to the transition app in the dev docs up to a point in the
page where it is last used to consolidate it with other links that are
in that portion of the page.