Add Fix For Issue 1827
Some of the SSH commands in the Developer Documentation are broken.
The common theme seems to be that they all have a different puppet_name
and app_name. The fix ensures the code checks for both variables before
rendering an error.

More information on this issue can be found here:
https://github.com/alphagov/govuk-developer-docs/issues/1827