Add transition checklist item to check for HTTPS
When transition was originally started, fewer sites were HTTPS enabled.
However, recently, sites to be transitioned are already served by
default over HTTPS. This is a problem since it would be a considerable
amount of work to regenerate our TLS certificate to include new domains
in our SAN everytime a site transitioned.

We should offer advice and guidance to departments who's domains are
not in our SAN when transitioning sites that are TLS terminated. In
these cases, setting up redirects on their existing infrastructure would
provide the most consistent service for now.