Create documentation for local replication of data with Docker
We developed a means to replicate data easily into a Docker container
for use with Whitehall, which is especially beneficial due to the
weight of the tool and amount of data present.

Before this commit there was no documentation on the process, and thus
this commit adds both this documentation and a relevant link from the
existing data replication page.