Revert autocorrected call to YAML.load
This break the parsing of the dates in the frontmatter and is a little
outside the scope of autocorrecting the linter.

Using safe_load results in the following error:

```
An error occurred while loading ./spec/pages/manual_page_spec.rb.
Failure/Error: frontmatter = YAML.safe_load(raw.split("---")[1])

Psych::DisallowedClass:
  Tried to load unspecified class: Date
 ./spec/pages/manual_page_spec.rb:4:in `block (2 levels) in <top (required)>'
 ./spec/pages/manual_page_spec.rb:2:in `block in <top (required)>'
 ./spec/pages/manual_page_spec.rb:1:in `each'
 ./spec/pages/manual_page_spec.rb:1:in `<top (required)>'
```