Add docs for the state of service resolution during migration
Due to the app by app strategy of the migration into AWS, the current
service discovery may be configured with significant differences
reflecting the state of the migration of an app as well as its
dependencies.
We add rough documentation on the edge cases to raise awareness that
changes may not be as straightforward as expected by developers.