Review High Priority Tests documentation
This adds a recently seen error which can be fixed using the same commands as
for the 'Traceback' error. There's also some repeition of killing the existing
Java process which doesn't seem necessary - having just worked through this
issue it seems doing so in conjunction with stopping and starting the
smokey-loop process does the trick.