Correct steps for leavers process
If we merge the second set of puppet PRs (that remove the user manifest
file) before we've removed the user from govuk-secrets, we risk errors
during a puppet deploy.

Rejig the ordering so that removing the user from govuk-secrets comes
before removing them from puppet. This makes it more explicit and will
encourage us to merge the secrets PR before the puppet PR.