Move govukcli setup instructions to get-ssh-access.html.md
It was being referred to in 'howto-ssh-to-machines-in-aws.html.md', which
is not in the standard flow of setup instructions for new joiners. Other
instructions expect govukcli to be set up already, which tripped up our
new starter.
We now move the setup instructions closer to the beginning of the
onboarding journey so that the step cannot be missed.