Tweak manual on deploying Terraform
This tweaks (actually, it's more of a rewrite) the manual around
deploying Terraform projects to be a bit more concise and clearly
highlight the steps to take, with less emphasis on explanation.

This also adds more sections to distinguish between deploing directly
from Jenkins vs. triggering a Jenkins job locally, which requires
a bit more effort to setup for small one-time changes.