Fix "run rake task" links for apps in AWS.
These used to link to Carrenza Jenkins even for apps in AWS, which was
exceedingly unhelpful.

The `blue.` in the Jenkins domain names is heinous but is necessary as a
workaround so that the deep links work, until we fix the GitHub OAuth
config for Jenkins to redirect to the name without the stackname.