Link to exact commit rather than master branch
Links with line numbers often fall out of date by tagging this to a
commit we can ensure the link remains working. It does risk that the
information could be out of date, but then the same could be the case
for this documentation if it links to out of date information.