Review logging
This removes the Elasticsearch section which seems to have got rather
confused in it's relation to Logit - I'm not too sure what this section
is saying anymore and thought it was easiest to remove.

It moves the Kibana section to be closer to the logit one given how
closely related they are.

It also removes the graphite section since this isn't a log.