Centralise alert docs for data sync
Previously the docs to support alerts for env-sync-and-backup were
mixed with the general documentation about the tool. The section about
the Icinga alert also didn't cover all the steps necessary to fix an
issue. This moves the parts of the general manual on env-sync-and-backup
to the existing docs for data-sync alerts, and modifies the text to
provide a clearer sequence of steps for 2ndline to follow.

Follow up work: modify Puppet to point env-sync-and-backup alerts to
the existing data-sync alert doc, instead of the general manual.