Rewrite AWS LB alert doc to highlight actions
Previously the alert doc for unhealthy AWS LBs was quite verbose and
contained quite a lot of repetition. This rewrites the doc to preserve
the same information and clearly highlight the steps to take in order
to debug the alert.

This also changes the owner Slack channel to 2ndline, since this is an
alert 2ndline may see and need to take action on.