Use fixture data in unit tests
I didn't like the unit tests being tied to the real data. For
example, if another repository is added to 'aws' that comes
before 'asset-manager' or after 'whitehall' in the alphabet,
the tests would start to fail. I also had no way of being
explicit about the apps returned from a 'apps_on_host' call.

I considered rewriting the tests programmatically, i.e. loop
through each returned result and check that the next app
in the list comes later in the alphabet. But this essentially
repeats the implementation in the test, which is not good
practice. It is better to rely on fixtures and to hardcode
the expected result.

There is a 'fixtures' directory I considered adding the
fixtures too, but this seemed an unnecessary pollutant when
the fixture data is relatively small and can live inline in
the test file itself. I suggest we only move the fixture
data into this folder if/when it can be used across multiple
tests.

A side effect of this refactor is the expected order of
hosts has changed ('ukcloud' coming before 'none', rather
than the other way around). I've left this in as it is
relatively unimportant - if multiple hosts have the same
count of repositories, their order is undefined.
In the next commit I will ensure that the hosts themselves
fall back to alphabetical order.