Return hosts in alphabetical order if repo count is the same
We should order by repo count where possible, but if two hosts
have the same repo count, we should order by the host name's
alphabetical position.

NB, this has allowed me to remove some of the fixture data from
earlier, which was only added to try to artificially influence
the order of the 'hosters_descending' return value.

'none' may be a special case we want to consider in the future
(for example, always appearing last regardless of repo count),
but that is outside the scope of this commit. For now it is
treated as a host like all the others.