Link to architectural deep-dive from 'Getting started'
Creates a new "You're all done" section to round off the guide,
just to confirm the developer has finished setting everything up.
This seems a natural point to introduce them to the architecture.

I considered linking them to the architectural overview [1]
instead, but in my experience the diagram doesn't mean much until
it has been explained, so the deep-dive is a better, if more
daunting piece of documentation to link to.

[1]: https://docs.publishing.service.gov.uk/manual/architecture.html