Review page on receiving emails from staging
This edits this doc page to reflect the fact that it relates only to
public facing emails and not to publisher emails. In order to do so, I
changed the title and added a note at the top of the page.

I also:
 - changed the file name to reflect the new tile
 - created a new placeholder file that will (in a later PR) document the
   process to receive publisher emails from Email Alert API in
   integration and staging
 - added a link to the new file in the aforementioned note

All this should hopefully clarify the distinction between public facing
emails and publishers emails and the different processes to get them in
our non-production environments.

The part on what to change in puppet was still showing outdated
instructions that are not relevant anymore (since we migrated
email_alert_api to AWS), so I edited that part too.