Update diagram
Removes old diagram depicting SSH to IP directly via jumpbox.
This is no longer possible as we don't forward the host keys.

As well as correcting that, this adds a shortcut method of
querying the jumpbox, by appending a jumpbox command at the
end of the gds govuk connect command.