Remove old (broken) SSH-from-jumpbox instructions
It is no longer possible to SSH from the Jumpbox using gds-cli,
as gds-cli/govukconnect do not forward your host keys to the
jumpbox for security reasons.
I've removed the old commands and replaced them with the ones
we are now expected to run.

I've also removed the troubleshooting guidance around SSH config,
which we should no longer have on our machines.